Fix for instance type switching on older pg versions (#22)
AWS and/or Terraform have some interesting logic on the svls v2 config.

This change + the existing local's defaults make it possible to switch
instance types from serverless to serverful (:P) - this worked previously, but
not on older versions of postgres that didnt support serverless, they would fail
to provision due to a check for the serverless config block (which happens to be present).